Implementing Retrofit & Testing a placeholder call:
Identified the correct data-type to use in Call<?> and Callback<?> that would present my Json data properly... It was actually more obvious than I expected. JsonArray (the gson variant). I had attempted to use JSONArray and List<JSONObject> and the answer was so close behind!

Cleaned up imports as well.